Frustum culling

+++ run_on_coal/Utils/Utils.cpp
@@ -28,12 +28,12 @@ GLint CheckShader(GLuint f_shader)
     return l_params;
 }
 
-void GetScreenPosition(glm::vec3 &pos, glm::mat4 &proj, glm::mat4 &view, glm::ivec2 &window, glm::ivec2 &out)
+void GetScreenPosition(glm::vec3 &pos, glm::mat4 &proj, glm::mat4 &view, glm::ivec2 &window, glm::vec2 &out)
 {
     glm::vec4 l_result = proj*view*glm::vec4(pos,1.0f);
     l_result /= l_result.w;
-    out.x = window.x/2*(1.f+l_result.x);
-    out.y = window.y/2*(1.f+l_result.y);
+    out.x = static_cast<float>(window.x)/2*(1.f+l_result.x);
+    out.y = static_cast<float>(window.y)/2*(1.f+l_result.y);
 }
 
 int ReadEnumString(std::string &f_val, const std::string &f_enum)
